Tidy some query functionality into a manager
